Extract media asset persistence snapshots
