+++ apps/backend/src/mediaAssets/persistence/index.ts
@@ -1,0 +1,27 @@
+export {
+  assertMediaBlobMatchesInput,
+  assertMediaBlobMatchesMetadata,
+  findMediaBlobRowBySha256InExecutor,
+  upsertMediaBlobRowInExecutor,
+} from "./blobs";
+export {
+  MEDIA_ASSET_COLUMNS,
+  MEDIA_ASSET_JOIN_CLAUSE,
+  MEDIA_BLOB_COLUMNS,
+  mapMediaAssetRow,
+  mapMediaAssetWithBlobRow,
+  mapMediaBlobRow,
+  toIsoString,
+  toOptionalIsoString,
+  toSafeNumber,
+} from "./rows";
+export {
+  findMediaAssetRowForUpdateInExecutor,
+  normalizeMediaAssetMutationMetadata,
+  normalizeMediaAssetSnapshotInput,
+  upsertMediaAssetSnapshotInExecutor,
+  upsertMediaAssetSnapshotWithBlobNormalizationInExecutor,
+} from "./snapshots";
+export {
+  recordMediaAssetSyncChange,
+} from "./syncChanges";
